# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Recall API — personal video memory search engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /people/enrolled:
    get:
      operationId: listEnrolledPeople
      tags: [people]
      summary: List people enrolled for face-based person search
      responses:
        "200":
          description: Enrolled people
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnrolledPerson"
    post:
      operationId: enrollPerson
      tags: [people]
      summary: Enroll a person for face-based search from a reference photo
      description: >-
        Indexes the largest face in the reference photo into the AWS
        Rekognition face collection and stores the returned FaceId under the
        given name.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/EnrollPersonInput"
      responses:
        "201":
          description: Person enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrolledPerson"
        "400":
          description: Unusable photo (no clear face) or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "502":
          description: Face indexing failed upstream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "503":
          description: Face recognition is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /people/enrolled/{id}:
    patch:
      operationId: updateEnrolledPerson
      tags: [people]
      summary: Rename an enrolled person
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrolledPersonUpdate"
      responses:
        "200":
          description: Updated person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrolledPerson"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    delete:
      operationId: deleteEnrolledPerson
      tags: [people]
      summary: Remove an enrolled person (and their face from the collection)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /voice/transcribe:
    post:
      operationId: transcribeVoice
      tags: [voice]
      summary: Transcribe a recorded voice query to text
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/VoiceTranscribeInput"
      responses:
        "200":
          description: Transcription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionResult"
        "503":
          description: Voice not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /voice/tts:
    post:
      operationId: synthesizeSpeech
      tags: [voice]
      summary: Speak an assistant answer aloud
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TtsInput"
      responses:
        "200":
          description: MP3 audio, base64-encoded for easy playback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TtsResult"
        "503":
          description: Voice not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
components:
  schemas:
...
    ApiMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    EnrolledPerson:
      type: object
      required: [id, name, thumbnailUrl, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        thumbnailUrl:
          type: string
          description: Data-URL copy of the reference photo for display
        createdAt: { type: string }
    EnrollPersonInput:
      type: object
      required: [name, photo]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        photo:
          type: string
          format: binary
          description: Reference photo (JPEG/PNG) with one clear face
    EnrolledPersonUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
...
    VoiceTranscribeInput:
      type: object
      required: [audio]
      properties:
        audio:
          type: string
          format: binary
          description: Recorded audio (webm/ogg/mp4/wav), max ~15MB
    TranscriptionResult:
      type: object
      required: [text]
      properties:
        text: { type: string }
    TtsInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1200
    TtsResult:
      type: object
      required: [audioBase64, mimeType]
      properties:
        audioBase64:
          type: string
          description: Base64-encoded MP3 audio of the spoken answer
        mimeType: { type: string }