# lib/api-zod/src/generated/types/ttsInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Recall API — personal video memory search engine
 * OpenAPI spec version: 0.1.0
 */

export interface TtsInput {
  /**
     * @minLength 1
     * @maxLength 1200
     */
  text: string;
}